refactor(patterns): consolidate pattern grammar into a single source of truth (#844)
Extract the authoritative pattern set + priority order from extractCitations into orderedPatterns (src/patterns/grammar.ts); extractCitations and the tokenize default both consume it. Removes tokenize's dead, incomplete, differently-ordered default list that the main pipeline never used. Behavior-preserving (token stream + dedup identical; full suite green). Adds tests/patterns/grammarOrder.test.ts to guard the load-bearing order against accidental reordering. Closes #844.
