+++ src/patterns/index.ts
@@ -11,6 +11,7 @@ export * from "./casePatterns"
 export * from "./constitutionalPatterns"
 export * from "./docketPatterns"
 export * from "./federalRulePatterns"
+export * from "./grammar"
 export * from "./journalPatterns"
 export * from "./legislativeMaterialPatterns"
 export * from "./localOrdinancePatterns"
